refactor: consolidate PROVIDER_LABELS in @chroxy/store-core
Move the canonical PROVIDER_LABELS map, getProviderLabel, getProviderInfo,
and related types from their duplicate homes in packages/app and
packages/dashboard into packages/store-core/src/provider-labels.ts.

Both packages already depend on @chroxy/store-core, so no new dependencies
are introduced. The app and dashboard files become thin re-export shims so
all existing import paths continue to work unchanged.

Resolves the gemini/codex label divergence ("Google Gemini"/"OpenAI Codex"
in app vs "Gemini (CLI)"/"Codex (CLI)" in dashboard) by adopting the
CLI-style labels as the single canonical set.

Closes #2974
